lab2 Borovsky_Maksim

+++ Reports/Borovsky_Maksim/lab2/task1.js
@@ -3,21 +3,21 @@ function task1()
 while(true)
 {
 var number1 = prompt("Введите число №1:");
-if(!parseInt(number1))
+if(!parseFloat(number1))
 {
 alert("первый ввод - не число");
 break;
 }
 
 var number2 = prompt("Введите число №2:");
-if(!parseInt(number2))
+if(!parseFloat(number2))
 {
 alert("второй ввод - не число");
 break;
 }
 
-var a = parseInt(number1);
-var b = parseInt(number2); 
+var a = parseFloat(number1);
+var b = parseFloat(number2); 
 
 if(a < b)
 alert("Второе число больше");
